% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adam_datasets}
\alias{adam_datasets}
\alias{sdtm_datasets}
\alias{adam_variables}
\title{Specs}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 23 rows and 12 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 48 rows and 5 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1878 rows and 21 columns.
}
\usage{
adam_datasets

sdtm_datasets

adam_variables
}
\description{
Specs
}
\examples{
library(tibble)

adam_datasets

adam_variables

sdtm_datasets
}
\keyword{datasets}
