% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pharmadverse}
\alias{pharmadverse}
\title{A data frame of the documentation of pharmaverse data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 74 rows and 7 columns.
}
\usage{
pharmadverse
}
\description{
A data frame of the documentation of pharmaverse data
}
\examples{
pharmadverse
}
\keyword{datasets}
