% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmaversedata-package.R
\docType{package}
\name{pharmaversedata-package}
\alias{pharmaversedata}
\alias{pharmaversedata-package}
\title{pharmaversedata: A Universe of Pharma Datasets}
\description{
Easy access to 'Pharmaverse' data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://maurolepore.github.io/pharmaversedata/}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{maurolepore@gmail.com} (\href{https://orcid.org/0000-0002-1986-7988}{ORCID})

}
\keyword{internal}
