% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{specs}
\alias{specs}
\title{A list with data specifications}
\format{
An object of class \code{list} of length 3.
}
\usage{
specs
}
\description{
A list with data specifications
}
\examples{
str(specs)
specs$define
}
\keyword{datasets}
